requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
drf-yasg
django-filter
